Improve searching for exact (ish) matches (#34)
This checks whether the query string is contained in any of the matches (instead of checking equality) so that we can get the desired card more easily in cases where some cards are named similarly (see `Mollog` and `Ollo`).
Before this change, typing `Mollog the` would still return both fighters (thanks to peculiarities of fuzzy searching), but afterwards `Ollo` will be filtered out.